kv-canary: 2nd-round review polish (sentinel-row test, dtype nit, axis consistency)
Review (Opus) follow-ups not blocking but worth folding before any merge
downstream:

1. Add test_swa_endpoint_trailing_sentinel_row_yields_skip in
   test_self_unit_endpoint.py: regression for the SWA endpoint's host
   gather propagating LUT[-1] sentinel rows. The prior kernel-side
   LUT lookup had this case covered via test_full_to_swa_lut_sentinel_
   skips_entry (now deleted); the new host gather has the same
   semantics but the coverage was lost.

2. endpoint.py launch_per_forward: drop redundant `.to(torch.int32)`
   on the gather output (LUT is already int32). Add a docstring note
   pinning the cuda-graph allocator contract.

3. bench_helpers.EXTEND_LEN_AXIS: add 16384 — the fast subset already
   pinned (bs=32, prefix=16384, extend=16384) as a long-prefill stress
   case (carried in from pre-cleanup); axis now matches.

4. bench_helpers.naive_slot_copy_fn: rename shadowed `total` →
   `n_slots`.

+++ python/sglang/srt/kv_canary/endpoint.py
@@ -95,11 +95,13 @@ class CanaryEndpoint:
         )
         # SWA endpoints translate the per-token slot indices host-side before invoking the write kernel —
         # the kernel itself is SWA-agnostic and only honours "slot < 0 ⇒ skip". FULL endpoints pass
-        # fb_out_cache_loc through unchanged.
+        # fb_out_cache_loc through unchanged. Under cuda-graph capture the two implicit allocations
+        # (int64 cast of fb_out_cache_loc + the gather output) are routed through PyTorch's private
+        # mempool — addresses are pinned on first capture and reused on replay.
         if self.full_to_swa_index_mapping is not None:
             fb_out_cache_loc_for_canary = self.full_to_swa_index_mapping[
                 fb_out_cache_loc.to(torch.int64)
-            ].to(torch.int32)
+            ]
         else:
             fb_out_cache_loc_for_canary = fb_out_cache_loc
